# Optimize `inline_memcpy` by `avx2`