--- a/libs/libcommon/CMakeLists.txt
+++ b/libs/libcommon/CMakeLists.txt
@@ -157,42 +157,51 @@ if (ENABLE_TESTS)
 endif ()
 
 # Region for specialized CPU flags tuning
-if (TIFLASH_ENABLE_AVX_SUPPORT)
-    # https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html
-    set_source_files_properties(src/mem_utils_avx2.cpp APPEND COMPILE_FLAGS "-mavx -mavx2")
-    set_source_files_properties(src/avx2_mem_utils_impl.cpp APPEND COMPILE_FLAGS "-mavx -mavx2 ${COMPILER_MOVBE_FLAG} -fomit-frame-pointer")
-    if (TIFLASH_COMPILER_VPCLMULQDQ_SUPPORT)
-        set_source_files_properties(src/crc64_avx2.cpp
-                APPEND COMPILE_FLAGS "-mavx2 -mpclmul -mvpclmulqdq -Wno-ignored-attributes")
-    endif()
-endif ()
 
-if (TIFLASH_ENABLE_AVX512_SUPPORT)
-    # https://gcc.gnu.org/onlinedocs/gcc/x86-Options.html
-    set_source_files_properties(src/mem_utils_avx512.cpp
-            APPEND COMPILE_FLAGS "-mavx512bw -mavx512vl -mavx512f")
-    if (TIFLASH_COMPILER_VPCLMULQDQ_SUPPORT)
-        set_source_files_properties(src/crc64_avx512.cpp
-                APPEND COMPILE_FLAGS "-mpclmul -mvpclmulqdq -mavx512dq -mavx512f -Wno-ignored-attributes")
-    endif()
-endif ()
+check_then_add_sources_compile_flag (
+    TIFLASH_ENABLE_AVX_SUPPORT
+    "-mavx2"
+    src/mem_utils_avx2.cpp
+    src/crc64_avx2.cpp
+    src/avx2_mem_utils_impl.cpp
+)
+check_then_add_sources_compile_flag (
+    TIFLASH_COMPILER_VPCLMULQDQ_SUPPORT
+    "-mvpclmulqdq;-Wno-ignored-attributes"
+    src/crc64_avx2.cpp
+    src/crc64_avx512.cpp
+)
+check_then_add_sources_compile_flag (
+    TIFLASH_COMPILER_MOVBE_SUPPORT
+    "-mmovbe" 
+    src/avx2_mem_utils_impl.cpp
+)
 
-if (TIFLASH_ENABLE_ASIMD_SUPPORT)
-    # https://gcc.gnu.org/onlinedocs/gcc/ARM-Options.html
-    set_source_files_properties(src/mem_utils_asimd.cpp APPEND COMPILE_FLAGS "-march=armv8-a+simd")
-    set_source_files_properties(src/crc64_sse2_asimd.cpp APPEND COMPILE_FLAGS "-march=armv8-a+simd+crypto")
-endif ()
+set_property(SOURCE src/avx2_mem_utils_impl.cpp APPEND PROPERTY COMPILE_OPTIONS "-fomit-frame-pointer") # remove frame pointer
+
+check_then_add_sources_compile_flag (
+    TIFLASH_ENABLE_AVX512_SUPPORT
+    "-mavx512bw;-mavx512vl;-mavx512f;-mavx512dq"
+    src/mem_utils_avx512.cpp
+    src/crc64_avx512.cpp
+)
+
+check_then_add_sources_compile_flag (
+    TIFLASH_ENABLE_ASIMD_SUPPORT
+    "-march=armv8-a+simd"
+    src/mem_utils_asimd.cpp
+)
+
+check_then_add_sources_compile_flag (
+    TIFLASH_ENABLE_ASIMD_SUPPORT
+    "-march=armv8-a+simd+crypto"
+    src/crc64_sse2_asimd.cpp
+)
 
 if (TIFLASH_ENABLE_SVE_SUPPORT)
     # https://gcc.gnu.org/onlinedocs/gcc/ARM-Options.html
     # set_source_files_properties(<SOURCE_FILE> APPEND COMPILE_FLAGS "-march=armv8.6-a+sve")
 endif ()
-
-if (ARCH_AMD64)
-    set_source_files_properties(
-            src/crc64_sse2_asimd.cpp
-            APPEND COMPILE_FLAGS "-mpclmul")
-endif()
 
 if (ARCH_AARCH64 AND OS_LINUX)
     target_link_libraries (common PUBLIC tiflash-aarch64-string tiflash-aarch64-math)