--- a/libs/libcommon/include/common/mem_utils_opt.h
+++ b/libs/libcommon/include/common/mem_utils_opt.h
@@ -22,13 +22,18 @@
 
 #if defined(__AVX2__)
 
-#define TIFLASH_USE_AVX2_COMPILE_FLAG 1
+constexpr bool tiflash_use_avx2_compile_flag = true;
 
 // if cpp source file is compiled with flag `-mavx2`, it's recommended to use inline function for better performance.
 #include <common/avx2_mem_utils.h>
 #include <common/avx2_strstr.h>
 
+#else
+constexpr bool tiflash_use_avx2_compile_flag = false;
 #endif
+
+#define ASSERT_USE_AVX2_COMPILE_FLAG \
+    static_assert(tiflash_use_avx2_compile_flag, __FILE__ " need compile flag `-mavx2`");
 
 #endif
 
@@ -69,7 +74,7 @@ FLATTEN_INLINE_PURE static inline bool IsStrViewEqual(const std::string_view & l
         return false;
 
 #ifdef TIFLASH_ENABLE_AVX_SUPPORT
-#ifdef TIFLASH_USE_AVX2_COMPILE_FLAG
+#ifdef __AVX2__
     return mem_utils::details::avx2_mem_equal(lhs.data(), rhs.data(), lhs.size());
 #else
     return mem_utils::avx2_mem_equal(lhs.data(), rhs.data(), lhs.size());
@@ -85,7 +90,7 @@ FLATTEN_INLINE_PURE static inline int CompareStrView(const std::string_view & lh
 #ifdef TIFLASH_ENABLE_AVX_SUPPORT
     const size_t size = std::min(lhs.size(), rhs.size());
 
-#ifdef TIFLASH_USE_AVX2_COMPILE_FLAG
+#ifdef __AVX2__
     int ret = mem_utils::details::avx2_mem_cmp(lhs.data(), rhs.data(), size);
 #else
     int ret = mem_utils::avx2_mem_cmp(lhs.data(), rhs.data(), size);
@@ -105,7 +110,7 @@ FLATTEN_INLINE_PURE static inline int CompareStrView(const std::string_view & lh
 FLATTEN_INLINE_PURE static inline size_t StrFind(std::string_view src, std::string_view needle)
 {
 #ifdef TIFLASH_ENABLE_AVX_SUPPORT
-#ifdef TIFLASH_USE_AVX2_COMPILE_FLAG
+#ifdef __AVX2__
     return mem_utils::details::avx2_strstr(src, needle);
 #else
     return mem_utils::avx2_strstr(src, needle);