--- a/libs/libcommon/include/common/memcpy.h
+++ b/libs/libcommon/include/common/memcpy.h
@@ -21,10 +21,15 @@
 #if defined(__SSE2__)
 #include <common/sse2_memcpy.h>
 #endif
+#if defined(__AVX2__)
+#include <common/avx2_memcpy.h>
+#endif
 
 ALWAYS_INLINE static inline void * inline_memcpy(void * __restrict dst, const void * __restrict src, size_t size)
 {
-#if defined(__SSE2__)
+#if defined(__AVX2__)
+    return mem_utils::avx2_inline_memcpy(dst, src, size);
+#elif defined(__SSE2__)
     return sse2_inline_memcpy(dst, src, size);
 #else
     return std::memcpy(dst, src, size);