--- a/libs/libmemcpy/memcpy.cpp
+++ b/libs/libmemcpy/memcpy.cpp
@@ -14,17 +14,13 @@
 
 #ifndef NO_TIFLASH_INTERNAL_MEMCPY
 
-#if defined(__SSE2__)
-
-#include <common/sse2_memcpy.h>
+#include <common/memcpy.h>
 
 /// This is needed to generate an object file for linking.
 
 extern "C" __attribute__((visibility("default"))) void * memcpy(void * __restrict dst, const void * __restrict src, size_t size)
 {
-    return sse2_inline_memcpy(dst, src, size);
+    return inline_memcpy(dst, src, size);
 }
-
-#endif
 
 #endif