@solotzg 2022-11-09
> LGTM Will we do similar optimization for `memcpy`?

The default value of `USE_INTERNAL_MEMCPY` is `ON` under x86-64 platform.

https://github.com/pingcap/tiflash/blob/40ad20c9439b18221297ed2b87d9d7421d17b174/libs/libmemcpy/CMakeLists.txt#L31-L42

The global memcpy function will be overrided by implementation here.
https://github.com/pingcap/tiflash/blob/dfac6a50b09711d12a3b6909126782d1b8ed363d/libs/libmemcpy/memcpy.cpp#L23