@Lloyd-Pottiger 2022-11-09
sorry, typo…… `memcmp`