@solotzg 2022-11-09
> > LGTM Will we do similar optimization for `memcpy`?
> > sorry, typo…… `memcmp`
> 
> By now, the comparison func of `StringRef` has been replaced with ours for most known cases(only need 3 options positive,negative,zero). But for memcmp, I'm not sure whether other modules rely on its return value.

Maybe that's not a problem and I'm overthinking.