@Lloyd-Pottiger 2022-11-09
> > so, the best practice for now is to replace `memcmp` with `mem_utils::IsStrViewEqual` / `mem_utils::CompareStrView`?
> 
> I think so. Using internal code can help compiler to inline it.

get it, thank you